requirements.txt
torch
Pillow
opencv-python
numpy
tqdm